Toilet facing direction code simplified (and error fixed)
Replaced a reference to a custom property from the toilet object, to the built in "getFacing" function. It returns the same IsoDirections object as the player, so I could easily reduce code by a ton.

In 42.13, checking the custom property was also causing an error, so two birds with one stone.